The way mouse entered and exited events are distributed have changed. Before mouse entered events were only distributed to the widget directy below the mouse and mouse exited events were distributed to the last widget whom had received a mouse entered event. That solution had a problem. Consider the case of a button inside a container. If the mouse would enter the container, the container would receive a mouse entered event. So far as expected, but if the mouse on a later stage would entered the button the container would receive a mouse exited - even though the mouse was still in the container area - and the button would receive a mouse entered event. One would expect the container to receive a mouse exited event only when the mouse actually leaves the container area.
The problem has been fixed using a stack holding all widgets with the mouse.
When the mouse enters a widget the widget and all of the widget's parents will
be pushed onto the stack (if they are not already present on the stack) and 
they will all receive a mouse entered event. As soon as the mouse leaves a 
widget present on the stack that widget will receive a mouse exited event
and it will be removed from the stack.